Move labeling image viewer to new component
